Power_sets in Python
I have used bitwise operators in this commit. Bitwise operators are operators that work on multi-bit values, but conceptually one bit at a time.
Shifts Left-wise
x << y
Returns x with the bits shifted to the left by y places (and new bits on the right-hand-side are zeros). This is the same as multiplying x by 2**y.

Shifts right-wise
x >> y
Returns x with the bits shifted to the right by y places. This is the same as //ing x by 2**y.

I have also used an alternative solution involving combinations and chain.from_iterable. Details about both the methods below:

itertools.combinations(iterable, r)
Return r length subsequences of elements from the input iterable.
P.S.: Combinations are emitted in lexicographic sort order. So, if the input iterable is sorted, the combination tuples will be produced in sorted order. Elements are treated as unique based on their position, not on their value. So if the input elements are unique, there will be no repeat values in each combination.

classmethod chain.from_iterable(iterable)
Alternate constructor for chain(). Gets chained inputs from a single iterable argument that is evaluated lazily.

What is chain used for?
itertools.chain(*iterables)
Make an iterator that returns elements from the first iterable until it is exhausted, then proceeds to the next iterable, until all of the iterables are exhausted. Used for treating consecutive sequences as a single sequence.