standard deviation
Objective 
In this challenge, we practice calculating standard deviation. 

Task 
Given an array, X, of N integers, calculate and print the standard deviation. Your answer should be in decimal form, rounded to a scale of 1 decimal place (i.e.,12.3 format). An error margin of +-0.1 will be tolerated for the standard deviation.

Input Format
The first line contains an integer, N, denoting the number of elements in the array. 
The second line contains N space-separated integers describing the respective elements of the array.