Basic Probability Puzzles
Objective 
In this challenge, we practice calculating probability.

Task 
There are 10 people about to sit down around a round table. Find the probability that 2 particular people will sit next to one another.

Output Format
In the editor below, submit your answer as Plain Text in the form of an irreducible fraction A/B, where A and B are both integers.