Only one person selected for a job
Bill and Nina appear for an interview for two vacancies. The probability of Bill’s selection is 1/3 and that of Nina’s selection is 1/5. Find the probability that only one of them will be selected.

Fill in the answer in the text box, in the form of an irreducible fraction A/B where A and B are two integers.