Find slope of regression line _ r-squared
Task 
There are 2 series of data involving index numbers: P for price index and S for the commodity stock. 
The mean and standard deviation of P are  100 and 8 respectively.
The mean and standard deviation of S are  103 and 4 respectively. 
The  R^2 correlation coefficient between the two series is 0.4.

With this data, obtain the slope of the regression line of  P on S, correct to a scale of 2 decimal places.

Output Format
Your answer should be a single floating point/decimal number, correct to a scale of 2 decimal places. You can submit solutions in either of the  following ways:

Solve the problem manually and submit your result as Plain Text. In the text box below, enter a single floating point/decimal number.
Submit an R or Python program, which uses the above parameters (hard-coded), then computes and prints the score.