Antony and the Magical Pearls
Sample Input
7  
2 4 1 2 10 2 3

Sample Output
59

Explanation
The cost of the seven pearls are 2,4,1,2,10,2,3 respectively. 
Initially, total cost=0. The bolded numbers are the pearls which Antony picks to combine. 
First, Antony picks the 1st and 3rd pearl, so combining cost=1+2=3 
Now, the cost of six pearls are 3,4,2,10,2,3 where the 1st pearl is the combination of the previous 1st and 3rd pearls. 
Next, Antony picks the 3rd and 5th pearl so combining cost=2+2=4 
Now, the cost of five pearls are 3,4,4,10,3 where the 3rd pearl is a combination of the previous 3rd and 5th pearls.
Next, Antony picks the 1st and and 5th pearl so combining cost=3+3=6 
Now, the cost of four pearls are 6,4,4,10 where the 1st pearl is the combination of the previous 1st and 5th pearls. 
Next, Antony picks the 2nd and 3rd pearl so combining cost=4+4=8 
Now, the cost of three pearls are 6,8,10 where the 2nd pearl is the combination of the previous 2nd and 3rd pearl. 
Next, Antony picks the 1st and 2nd pearl so combining cost=8+6=14 
Now, the cost of two pearls are 14,10 where the 1st pearl is a combination of the previous 1st and 2nd pearls. 
Finally, Antony picks the last two pearls so combining cost=14+10=24 

Hence, the total cost for the Nth pearl =  Sigma(merging cost) = 3+4+6+8+14+24=59 
You can check and verify that picking pearls in any other order will not give the minimum cost for the final pearl.